Type change
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3818 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ncarview/src/lib/libctrans/Xdefs.h
@@ -37,7 +37,7 @@ Colormap	Cmap;		/* current colormap for $win	*/
 int		ColorModel;	/* color model			*/
 boolean		MyCmap;		/* I created Cmap		*/
 boolean		RoCmap;		/* Read Only Cmap		*/
-int		ColorErr;	/* % color error allowed	*/
+float		ColorErr;	/* % color error allowed	*/
 /*
  *      If true the driver will not attempt to set the background
  *      color or clear the window.
